+++ changelog.md
@@ -8,11 +8,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- `batchapi` module ([Issue #49](https://github.com/michaeldcanady/servicenow-sdk-go/issues/49))
+- `tableapi.TableItemRequestBuilder2`
+- `tableapi.NewTableItemRequestBuilder2`
+- `tableapi.TableRequestBuilder2`
+- `tableapi.NewTableRequestBuilder2`
+- `servicenowsdk.NowRequestBuilder.Table2`
+
 ### Changed
 
 ### Deprecated
 
+- `tableapi.TableItemRequestBuilder`
+- `tableapi.NewTableItemRequestBuilder`
+- `tableapi.TableRequestBuilder`
+- `tableapi.NewTableRequestBuilder`
+- `servicenowsdk.NowRequestBuilder.Table`
+
 ### Removed
+
+### Fixed
+
+## [1.6.1] - (09/03/2024)
 
 ### Fixed